+++ apps/mobile/utils/changelogData.ts
@@ -9,27 +9,41 @@
 /**
  * Raw CHANGELOG.md content embedded at build time.
  */
-export const CHANGELOG_TEXT = `## [1.0.76](https://github.com/alternun-development/alternun/compare/v1.0.75...v1.0.76) (2026-04-15)
-
+export const CHANGELOG_TEXT = `## [1.0.78](https://github.com/alternun-development/alternun/compare/v1.0.77...v1.0.78) (2026-04-15)
 
 ### Bug Fixes
 
-* **repo:** sync auth exports and footer drawers ([74f4d70](https://github.com/alternun-development/alternun/commit/74f4d702be16695f1738a9422282cc4d7dcc9b56))
+- **docs,mobile,auth:** CHANGELOG, PublicLandingPage, AirsIntroExperience, changelogData ([83934d7](https://github.com/alternun-development/alternun/commit/83934d7d35564e3ee578c10ea265b0b6f1667a74))
 
+## [1.0.78](https://github.com/alternun-development/alternun/compare/v1.0.77...v1.0.78) (2026-04-15)
 
+### Bug Fixes
 
+- **docs,mobile,auth:** CHANGELOG, PublicLandingPage, AirsIntroExperience, changelogData ([83934d7](https://github.com/alternun-development/alternun/commit/83934d7d35564e3ee578c10ea265b0b6f1667a74))
 
+## [1.0.77](https://github.com/alternun-development/alternun/compare/v1.0.76...v1.0.77) (2026-04-15)
+
+### Bug Fixes
+
+- **auth:** remove server-only and better-auth exports from main index ([4e391d1](https://github.com/alternun-development/alternun/commit/4e391d1c161e3f7196662e78d04a73ad27057192))
+
+## [1.0.77](https://github.com/alternun-development/alternun/compare/v1.0.76...v1.0.77) (2026-04-15)
+
+### Bug Fixes
+
+- **auth:** remove server-only and better-auth exports from main index ([4e391d1](https://github.com/alternun-development/alternun/commit/4e391d1c161e3f7196662e78d04a73ad27057192))
 
 ## [1.0.76](https://github.com/alternun-development/alternun/compare/v1.0.75...v1.0.76) (2026-04-15)
 
+### Bug Fixes
+
+- **repo:** sync auth exports and footer drawers ([74f4d70](https://github.com/alternun-development/alternun/commit/74f4d702be16695f1738a9422282cc4d7dcc9b56))
+
+## [1.0.76](https://github.com/alternun-development/alternun/compare/v1.0.75...v1.0.76) (2026-04-15)
 
 ### Bug Fixes
 
-* **repo:** sync auth exports and footer drawers ([74f4d70](https://github.com/alternun-development/alternun/commit/74f4d702be16695f1738a9422282cc4d7dcc9b56))
-
-
-
-
+- **repo:** sync auth exports and footer drawers ([74f4d70](https://github.com/alternun-development/alternun/commit/74f4d702be16695f1738a9422282cc4d7dcc9b56))
 
 ## [1.0.75](https://github.com/alternun-development/alternun/compare/v1.0.74...v1.0.75) (2026-04-15)
 
@@ -1372,4 +1386,4 @@ All notable changes to this project will be documented in this file.
 /**
  * Version extracted from package.json for display purposes.
  */
-export const APP_VERSION = '1.0.76';
+export const APP_VERSION = '1.0.78';